started working on python version

+++ include/voxelize_cpu.hpp
@@ -5,7 +5,7 @@ namespace Voxelize {
 // TODO
 // documentation
 void
-voxelize(uint64_t Nparticles, int64_t box_N, int64_t dim, float box_L,
+voxelize(size_t Nparticles, size_t box_N, size_t dim, float box_L,
          float *coords, float *radii, float *field, float *box);
 
 } // namespace Voxelize
