+++ include/voxelize_gpu.hpp
@@ -1,4 +1,4 @@
-#include <cstdint>
+#include <cstddef>
 
 #include "gpu_handler.hpp"
 
@@ -7,7 +7,7 @@ namespace Voxelize {
 // TODO
 // documentation
 void
-voxelize(uint64_t Nparticles, int64_t box_N, int64_t dim, float box_L,
+voxelize(size_t Nparticles, size_t box_N, size_t dim, float box_L,
          float *coords, float *radii, float *field, float *box,
          gpu_handler * gpu);
